actionpack#117: "adds more remove tests" (#125)
* removes comments

* adds tests of Remove from empty file

* updates Remove.instruction

* adds tests of Remove from head and tail

+++ actionpack/actions/remove.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 from pathlib import Path
-from re import finditer
 from typing import Union
 
 from actionpack import Action
@@ -17,17 +16,15 @@ class Remove(Action[Name, FileContents]):
         self.tail = tail
 
     def instruction(self) -> FileContents:
-        # (line.group(0) for line in finditer(r'\w.+', self.path.read_text()))
         with open(str(self.path), 'r+') as file:
             lines = file.readlines()
-            to_remove = lines.pop(-1) if self.tail else lines.pop(0)
-            file.seek(0)
-            for line in lines:
-                file.write(line)
-            file.truncate()
-        return to_remove
-        # return to_remove.decode('') if self.output_type is str else to_remove.encode()
-        # return self.path.read_bytes() if self.output_type is bytes else self.path.read_text()
+            if lines:
+                to_remove = lines.pop(-1) if self.tail else lines.pop(0)
+                file.seek(0)
+                for line in lines:
+                    file.write(line)
+                file.truncate()
+        return to_remove if lines else ''
 
     def validate(self) -> Remove[Name, FileContents]:
         if not self.path.exists():
